requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
mysql-connector-python==8.0.15
protobuf==3.6.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
